Setup Github actions for use in SQM projects
This replaces the Github actions used by the official Bootstrap repo
with ones that allow us to maintain and publish the gem in the SQM
package repo. The actions are the same as the ones we use for our own
gems, with the following small differences:

1. Bootstrap uses "main" instead of "master" for its release branch.
2. Using Bootstrap's Rake task for building the gem.
3. Releasing the packaged gem from the `pkg` directory.

This should ensure that the modified gem that we use is a close as
possible to the official Bootstrap release.

This also updates the gem's version number, combining the minor and
point release numbers as the minor number and using the point release
number to designate the custom SQM version (e.g. "5.3.3" becomes
"5.33.1"). This is done because the version-forget-me-not Github action
can only support semantic versioning with three numbers.

[1207966452799142](https://app.asana.com/0/0/1207966452799142/f)